Include a length validation for Question description
We really need very descriptive questions to the students answer. A problem I had in my master degree course, in the last experiment, was poor questions descriptions, that doesn't contain specific details about the program execution, making the students have a hard work. So I enforced a minimum length for the description as a try to avoid this.