Implement TestCasePolicy
TestCasePolicy is very similar to QuestionPolicy, so much that is its parent class. The differences are:

1) index, create and destroy already will check if the test case question belongs to the current user;
2) We have a specific rule for destroy, where cannot be performed if the test case question is already present in and exercise and have only this test case implemented.

The reason of difference 2 is because that can causes inconsistency in the answer process of an exercise with the question of test case: once it haven't any test case implemented, any answer submited will be "correct".