Introduce ClassRoom module and create EnrollmentController
To provide a better organization, I'll start to concentrate some main functionalities inside the module ClassRoom.
This module will concentrate everithing regarding a class room: enrollments, inclusion of exercises to the team, question's answerings, etc.
As a start for this module, this commit implements the EnrollmentController, providing a way to the users enroll and unenroll in a team.